+++ constants/route.tsx
@@ -1,4 +1,3 @@
-export const ALLPOKEMON: string = "ALLPOKEMON";
-export const DETAILPAGE: string = "DETAILPAGE";
 export const LISTPAGE:string = "LISTPAGE"
+export const DETAILPAGE: string = "DETAILPAGE";
 export const OWNEDPAGE:string = "OWNEDPAGE"
