+++ pages/index.tsx
@@ -5,14 +5,16 @@ import { css, jsx } from "@emotion/react";
 import type { NextPage } from "next";
 import Head from "next/head";
 
-import { FC, useState } from "react";
+import { FC, useState, Dispatch, SetStateAction } from "react";
+import { useQuery, gql } from "@apollo/client";
 
 import Layout from "@components/Layout";
 import Listpage from "@components/Listpage/Listpage";
+import Detailpage from "@components/Detailpage/Detailpage";
 import Ownedpage from "@components/Ownedpage/Ownedpage";
-import { LISTPAGE, OWNEDPAGE } from "@constants/route";
 import Navigator from "@components/Navigator";
 import Search from "@components/Search";
+import { DETAILPAGE, LISTPAGE, OWNEDPAGE } from "@constants/route";
 
 const color = "white";
 
@@ -22,48 +24,87 @@ const example = css`
   }
 `;
 
-const containerStyle = css`
-  padding: 0 0.5rem;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-  width: 100%;
-  overflow: hidden;
+const GET_POKEMONS = gql`
+  query pokemons($limit: Int, $offset: Int) {
+    pokemons(limit: $limit, offset: $offset) {
+      count
+      results {
+        id
+        name
+        image
+        artwork
+        dreamworld
+      }
+    }
+  }
 `;
 
-interface ContentPropsType {
-  page: string;
+interface IContentProps {
+  pokemons: any;
+  currentPage: string;
+  setCurrentPage: Dispatch<SetStateAction<string>>;
+  currentId: number;
+  setCurrentId: Dispatch<SetStateAction<number>>;
+  currentName: string;
+  setCurrentName: Dispatch<SetStateAction<string>>;
 }
 
-const Content: FC<ContentPropsType> = (props) => {
-  const { page } = props;
+const Content: FC<IContentProps> = (props) => {
+  const {
+    pokemons,
+    currentPage,
+    setCurrentPage,
+    currentId,
+    setCurrentId,
+    currentName,
+    setCurrentName,
+  } = props;
 
-  switch (page) {
+  switch (currentPage) {
     case LISTPAGE:
-      return <Listpage />;
+      return <Listpage pokemons={pokemons} setCurrentPage={setCurrentPage} setCurrentId={setCurrentId} setCurrentName={setCurrentName}/>;
 
     case OWNEDPAGE:
-      return <Ownedpage />;
+      return <Ownedpage setCurrentPage={setCurrentPage} setCurrentName={setCurrentName} />;
+
+    case DETAILPAGE:
+      return <Detailpage setCurrentPage={setCurrentPage} id={currentId} name={currentName} />;
+
     default:
-      return <></>;
+      return <div>ERROR</div>;
   }
 };
 
 const Home: NextPage = () => {
   const [currentPage, setCurrentPage] = useState<string>("LISTPAGE");
+  const [currentId, setCurrentId] = useState<number>(1);
+  const [currentName, setCurrentName] = useState<string>("");
+
+  const { loading, error, data } = useQuery(GET_POKEMONS, {
+    variables: { limit: 1000, offset: 0 },
+  });
+
+  if (loading) return <div>Loading</div>;
+  if (error) return <div>Error</div>;
 
   return (
     <div>
       <Head>
         <title>Pokemon Cool</title>
-        <meta name="description" content="Generated by create next app" />
+        <meta name="description" content="" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
 
       <Layout>
-        <Content page={currentPage} />
+        <Content
+          currentPage={currentPage}
+          setCurrentPage={setCurrentPage}
+          currentId={currentId}
+          setCurrentId={setCurrentId}
+          currentName={currentName}
+          setCurrentName={setCurrentName}
+          pokemons={data.pokemons}
+        />
         <Navigator currentPage={currentPage} setCurrentPage={setCurrentPage} />
         <Search />
       </Layout>
